"petis bugs sur l'api et la prise en compte de * sur #LESVALEURS*"
git-svn-id: svn://zone.spip.org/spip-zone/_plugins_@11627 ac52e18a-acf5-0310-9fe8-c4428f23b10a

+++ _stable_/forms/forms_et_tables_1_9_1/inc/forms.php
@@ -471,10 +471,12 @@
 					$inserts = array_merge($inserts,$ins);
 				}
 			}
-			$in_champs = calcul_mysql_in('champ',join(',',array_map('_q', $champs_mod)));
-			spip_query("DELETE FROM spip_forms_donnees_champs WHERE $in_champs AND id_donnee="._q($id_donnee));
-			spip_query($q="INSERT INTO spip_forms_donnees_champs (id_donnee, champ, valeur) ".
-				"VALUES ".join(',', $inserts));
+			if (count($inserts)){
+				$in_champs = calcul_mysql_in('champ',join(',',array_map('_q', $champs_mod)));
+				spip_query("DELETE FROM spip_forms_donnees_champs WHERE $in_champs AND id_donnee="._q($id_donnee));
+				spip_query($q="INSERT INTO spip_forms_donnees_champs (id_donnee, champ, valeur) ".
+					"VALUES ".join(',', $inserts));
+			}
 		}
 		else
 			spip_log("erreur: ".serialize($erreur));
